Implement onComputeColors() for live wallpaper for background-based dynamic system colors
On S and later, this color is used to dynamically switch the theme color
of the system UI. Implement this method for a nicer user experience.

Note that we do not use the colors defined in resources or the compose
theme directly, since those hard-code one set of theme colors for day or
night mode, but with live wallpapers we would like the system color to
match what we are actually currently displaying.