fix bug where trailing spaces would hide relevant results
When using the search and forgetting to remove trailing white spaces, you will not get the most relevant results displayed first. Breezy Weather should not care about whether the search contains a trailing white space or not, as this can happen very easily by using autocomplete on GBoard, for example.