IMS: fix missing daily high/low if installing after evening
the code calculated daily high/low from the hourly part of the API.
IMS does not send hourly history, so 17:00 the API will only have evening/night temperatures.
when installing the app after that time, the "high" temperature will not calculate.
IMS API does return daily min and max temps, so use it to determine the daily min/max temps in the app.